refactor(web): share page keyboard-shortcut hook
